Overhaul keymaps
This commit reduces the set of keymaps to only two files, one for each
dialect. These appear to be the most commonly used mappings and have the
advantage of minimizing unnecessary differences while providing full
coverage of the alphabet and standard punctuation marks. The old set of
mappings will be maintained in a separate branch.